move posts queries to entities
