+++ src/entities/Post/api/index.ts
@@ -1,0 +1,1 @@
+export { postsApiClient } from './postsApiClient';
