+++ src/entities/Post/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useGetPostsRequest, useGetPostRequest } from './requestHooks';
