+++ src/entities/Post/types/index.ts
