+++ src/entities/Post/types/types.ts
