+++ src/shared/api/index.ts
@@ -1,2 +1,1 @@
 export { httpClient } from './httpClient';
-export { postsApiClient } from './postsApiClient';
