+++ src/shared/hooks/index.ts
@@ -1,3 +1,2 @@
 export { useOnlineStatus } from './useOnlineStatus';
 export { useDebounce } from './useDebounce';
-export { useGetPostsRequest, useGetPostRequest } from './requestHooks';
