+++ src/shared/index.ts
@@ -3,4 +3,3 @@ export * from './api';
 export * from './config';
 export * from './ui';
 export * from './hooks';
-export * from './types';
